third_party/WebKit/LayoutTests/inspector/elements/styles/filter-matched-styles-expected.txt

Verifies that filtering in StylesSidebarPane works as expected.


Running: selectInitialNode

Running: testSimpleFiltering
Filtering styles by: padding
element.style {
}
.mydiv {
   border:>1px solid black
F  padding:v10px 10px 10px 10px
F      padding-top: 10px
F      padding-right: 10px
F      padding-bottom: 10px
F      padding-left: 10px
}

Running: testLonghandsAreAutoExpanded
Filtering styles by: -top
element.style {
}
.mydiv {
   border:v1px solid black
F      border-top-color: black
F      border-top-style: solid
F      border-top-width: 1px
       border-right-color: black
       border-right-style: solid
       border-right-width: 1px
       border-bottom-color: black
       border-bottom-style: solid
       border-bottom-width: 1px
       border-left-color: black
       border-left-style: solid
       border-left-width: 1px
   padding:v10px 10px 10px 10px
F      padding-top: 10px
       padding-right: 10px
       padding-bottom: 10px
       padding-left: 10px
}

Running: testAutoExpandedLonghandsAreCollapsed
Filtering styles by: null
element.style {
}
#inspected {
   border-size: 2px
}
.mydiv {
   border:>1px solid black
   padding:>10px 10px 10px 10px
}
div {
   display: block
}
